Enable singleton handling for more cases
