+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/internal/cc/impl/DefaultConfigurationCacheIO.kt
@@ -558,8 +558,7 @@ class DefaultConfigurationCacheIO internal constructor(
             flowProviders = service(),
             transformStepNodeFactory = service(),
             parallelStore = startParameter.isParallelStore,
-            parallelLoad = startParameter.isParallelLoad,
-            enableSingletonCaching = startParameter.enableSingletonCaching
+            parallelLoad = startParameter.isParallelLoad
         )
 
     private
