+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/internal/cc/impl/serialize/Codecs.kt
@@ -20,8 +20,6 @@ import org.gradle.api.file.FileSystemOperations
 import org.gradle.api.flow.FlowProviders
 import org.gradle.api.internal.DocumentationRegistry
 import org.gradle.api.internal.GradleInternal
-import org.gradle.api.internal.MutationGuards
-import org.gradle.api.internal.artifacts.DefaultBuildIdentifier
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ArtifactSetToFileCollectionFactory
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.result.CapabilitySerializer
 import org.gradle.api.internal.artifacts.transform.TransformActionScheme
@@ -39,13 +37,11 @@ import org.gradle.api.internal.file.collections.DirectoryFileTreeFactory
 import org.gradle.api.internal.provider.PropertyFactory
 import org.gradle.api.internal.provider.ValueSourceProviderFactory
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.specs.Specs
 import org.gradle.api.tasks.util.PatternSet
 import org.gradle.composite.internal.BuildTreeWorkGraphController
 import org.gradle.execution.plan.OrdinalGroupFactory
 import org.gradle.execution.plan.TaskNodeFactory
 import org.gradle.internal.Factory
-import org.gradle.internal.ImmutableActionSet
 import org.gradle.internal.build.BuildStateRegistry
 import org.gradle.internal.execution.InputFingerprinter
 import org.gradle.internal.hash.ClassLoaderHierarchyHasher
@@ -141,7 +137,6 @@ import org.gradle.internal.serialize.graph.codecs.NotImplementedCodec
 import org.gradle.internal.serialize.graph.codecs.ServicesCodec
 import org.gradle.internal.serialize.graph.reentrant
 import org.gradle.internal.state.ManagedFactoryRegistry
-import org.gradle.util.Path
 
 
 @Suppress("LongParameterList")
@@ -179,8 +174,7 @@ class Codecs(
     flowProviders: FlowProviders,
     transformStepNodeFactory: TransformStepNodeFactory,
     val parallelStore: Boolean = true,
-    val parallelLoad: Boolean = true,
-    val enableSingletonCaching: Boolean = true
+    val parallelLoad: Boolean = true
 ) {
     private
     val userTypesBindings: Bindings
@@ -193,10 +187,6 @@ class Codecs(
             unsupportedTypes()
 
             baseTypes()
-
-            if (enableSingletonCaching) {
-                addSingletons()
-            }
 
             bind(HASHCODE_SERIALIZER)
 
@@ -397,13 +387,4 @@ class Codecs(
 
     fun workNodeCodecFor(gradle: GradleInternal, contextSource: IsolateContextSource) =
         WorkNodeCodec(gradle, internalTypesCodec(), ordinalGroupFactory, contextSource, parallelStore, parallelLoad)
-}
-
-private fun BindingsBuilder.addSingletons() {
-    withSingleton(Path.ROOT)
-    withSingleton(DefaultBuildIdentifier.ROOT)
-    withSingleton(Specs.SATISFIES_ALL)
-    withSingleton(Specs.SATISFIES_NONE)
-    withSingleton(ImmutableActionSet.empty<Any>())
-    withSingleton(MutationGuards.identity())
 }
